test(crew): bring the render-asserting files in step with the spend grid
The wider fixed grid leaves the reason fewer columns, so an assertion
that rendered a full clause at the default width truncated once the
spend block landed; those tests now render at the widened pane width the
workstation measures, where the clause still fits in full. The same pass
retires the pre-existing drift those files carried against the current
renderer: counter separators without surrounding spaces, a completed
manifest without an independent review read as completed_unpromoted, and
a wait fixture whose start was frozen to a date that the suite's own run
day eventually turned overdue. A test must not encode the current date.

+++ reckon/crew/ticker.py
@@ -388,7 +388,7 @@ def local_clock(observed: Any) -> str:
 
 def _clock(seconds: float) -> str:
     """Render a duration as h:mm:ss (or m:ss under an hour), seconds rounded."""
-    total = max(0, int(round(seconds)))
+    total = max(0, round(seconds))
     hours, remainder = divmod(total, 3600)
     minutes, secs = divmod(remainder, 60)
     if hours:
@@ -716,6 +716,7 @@ class Ticker:
         measured zero stays a zero. The cells read the record's own figures and
         shape them only here, so the persisted event stays re-renderable.
         """
+
         def cell(value: Any, formatter: Callable[[float], str]) -> tuple[str, Any]:
             if isinstance(value, Real):
                 return formatter(float(value)), None
